requirements.txt
streamlit>=1.31.0
langgraph>=0.2.0
langchain>=0.2.0
langchain-google-genai>=1.0.0
google-generativeai>=0.3.0
pillow>=10.0.0
pypdf>=4.0.0
python-dotenv>=1.0.0
pydantic>=2.0.0
typing-extensions>=4.8.0

